Parameter Spaces

TODO

Methods

Public methods

ParameterSpaces$new()

ParameterSpaces$print()

ParameterSpaces$switch()

ParameterSpaces$o2i()

ParameterSpaces$i2o()

ParameterSpaces$clone()

Method new()

Initializes a new ParameterSpaces object.

Usage

ParameterSpaces$new(parameter_names, parameter_lengths_in_o_space)

Arguments

parameter_names

[character()]
Unique names for the parameters.

parameter_lengths_in_o_space

[integer()]
The length of each parameter in the optimization space.

Returns

A new ParameterSpaces object.

Method print()

Print an overview of the parameter spaces.

Usage

ParameterSpaces$print(show_transformer = FALSE)

Arguments

show_transformer

[logical(1)]
Show transformer functions in the output?

Method switch()

Switch between optimization and interpretation space.

Usage

ParameterSpaces$switch(x)

Arguments

x

[numeric() | list()]
The parameters, either as numeric() (will be switched to interpretation space), or as list() (will be switched to optimization space).

Method o2i()

Define transformation functions when switching from Optimization Space to Interpretation Space.

Usage

ParameterSpaces$o2i(...)

Arguments

...

[function]
One or more transformation functions, named according to the parameters.

Transformers from Optimization Space to Interpretation Space (o2i) must receive a numeric. The default is the identity.

Method i2o()

Define transformation functions when switching from Interpretation Space to Optimization Space.

Usage

ParameterSpaces$i2o(...)

Arguments

...

[function]
One or more transformers functions, named according to the parameters.

Transformers from Interpretation Space to Optimization Space (i2o) must return a numeric. The default is as.vector().

Method clone()

The objects of this class are cloneable with this method.

Usage

ParameterSpaces$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# TODO