Apply optimizer object

This function performs numerical optimization using an optimizer object.

apply_optimizer(optimizer = optimizer_nlm(), objective, initial, ...)

Arguments

optimizer

An object of class optimizer.

objective

The function to be optimized, returning a single numeric. Its first argument must be a numeric vector of the length of initial, followed by any other arguments specified by the ... argument.

initial

A numeric vector with starting parameter values for the optimization.

...

Additional arguments to be passed to objective.

Value

A named list, containing at least three elements:

value

A numeric, the value of the estimated optimum of objective.

parameter

A numeric vector, the parameter vector where the optimum of objective is obtained.

secondsA numeric, the total optimization time in seconds.

Additional output elements of the optimizer (if not excluded by the output_ignore element via set_optimizer) are appended.

See also

Examples

apply_optimizer(optimizer_nlm(), function(x) -(x + 1)^2 + 1, 2)
#> $v
#> [1] -148523066
#> 
#> $z
#> [1] 12186
#> 
#> $time
#> Time difference of 0.01068497 secs
#> 
#> $gradient
#> [1] -24374.02
#> 
#> $code
#> [1] 5
#> 
#> $iterations
#> [1] 11
#>